split motor from canbus
